+++ crates/lib/src/action/outcome_worker.rs
@@ -9,6 +9,7 @@ use futures::Future;
 use tokio::{
 	spawn,
 	sync::{mpsc, watch::Receiver},
+	time::sleep,
 };
 use tracing::{debug, error, trace, warn};
 
@@ -151,6 +152,10 @@ impl OutcomeWorker {
 
 			(true, Outcome::Wait) => {
 				notry!(self.process.wait())?;
+			}
+
+			(_, Outcome::Sleep(time)) => {
+				notry!(sleep(time));
 			}
 
 			(_, Outcome::Clear) => {